Update SendGrid comment and add SMTP timeout config
The comment regarding sending emails via SendGrid has been updated to include a link to obtain an API key and to note that SMTP settings cannot be used with SendGrid as of December 2020.

A new configuration key `SmtpTimeoutMs` with a value of `5000` has been added to specify the SMTP timeout in milliseconds.